src/backend/requirements.txt
Flask
Flask-SocketIO
python-dotenv
web3
google-generativeai
supabase
requests
Pillow
gradio
python-socketio[client]
eventlet